requirements.txt
absl-py==0.9.0
astor==0.8.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-RESTful==0.3.7
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gast==0.2.2
google-pasta==0.1.8
grpcio==1.26.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
numpy==1.16.2
opencv-python-headless
opencv-contrib-python-headless
opt-einsum==3.1.0
pafy==0.5.5
Pillow==6.2.0
protobuf==3.11.2
pyparsing==2.4.6
python-dateutil==2.8.1
requests==2.22.0
scikit-learn==0.20.2
scipy==1.2.1
six==1.14.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.25.8
Werkzeug==0.16.1
wrapt==1.11.2
youtube-dl==2020.1.24